rename conflict file name
